Refactor - Extract static nested classes into their own class
